Move `basic_object`to parent directory
