+++ NitrotypePy/objects/api/info/bootstrap/basic_object.py
