+++ NitrotypePy/objects/api/info/bootstrap/item/loot.py
@@ -1,4 +1,4 @@
-from .basic_object import Options
+from ..basic_object import Options
 from typing import TypedDict
